# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/card": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создать карту со свойствами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Создать карту",
                "parameters": [
                    {
                        "description": "Card data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateCardRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemas.CreateCardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/card/change_pic/{ID}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Поменять картинку используя ее ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Поменять картинку по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Picture was changed sucessfully",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/card/{ID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить информацию о карте по ее названию",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Получить карту по названию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetCardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновить карту по ее ID с параметрами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Обновить карту по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update card data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.UpdateCardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удаляет карту по ее ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Удалить карту по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.DeleteCardResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/card_to_move/{ID}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Этот эндпойнт позволяет добавить карту в ход по ее ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Добавить карту в ход",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Card ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Card added successfully",
                        "schema": {
                            "$ref": "#/definitions/schemas.AddCardToMoveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/cards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список всех карт.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Получить все карты",
                "responses": {
                    "200": {
                        "description": "List of cards retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetAllCardsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.Cards": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "long_description": {
                    "type": "string"
                },
                "multiplier": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "title_en": {
                    "type": "string"
                },
                "title_ru": {
                    "type": "string"
                }
            }
        },
...
        "schemas.AddCardToMoveResponse": {
            "type": "object",
            "properties": {
                "cardID": {
                    "type": "integer"
                },
                "messageResponse": {
                    "type": "string"
                },
                "moveID": {
                    "type": "integer"
                }
            }
        },
...
        "schemas.CreateCardRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "long_description": {
                    "type": "string"
                },
                "multiplier": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "title_en": {
                    "type": "string"
                },
                "title_ru": {
                    "type": "string"
                }
            }
        },
        "schemas.CreateCardResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "messageResponse": {
                    "type": "string"
                }
            }
        },
...
        "schemas.DeleteCardResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "messageResponse": {
                    "type": "string"
                }
            }
        },
        "schemas.GetAllCardsResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Cards"
                    }
                },
                "count": {
                    "type": "integer"
                },
                "move_ID": {
                    "type": "integer"
                }
            }
        },
...
        "schemas.GetCardResponse": {
            "type": "object",
            "properties": {
                "card": {
                    "$ref": "#/definitions/ds.Cards"
                }
            }
        },
...
        "schemas.ResponseMessage": {
            "type": "object"
        },
        "schemas.UpdateCardRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "long_description": {
                    "type": "string"
                },
                "multiplier": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "title_en": {
                    "type": "string"
                },
                "title_ru": {
                    "type": "string"
                }
            }
        },
        "schemas.UpdateCardResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "messageResponse": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}